Add EventFactory service, SVG files, and Prisma schema configurations
This commit includes several additions to the project to support various functionalities. It introduces a new EventFactory service responsible for event creation. It also adds SVG files for UI elements. Moreover, Prisma schema configurations have been created to define the structure and relationships of database entities in the application. Other minor changes and improvements are applied as well.